FEATURE: servo positiion is now a float and can take decimal values

+++ Drone/tracker.h
@@ -31,7 +31,7 @@ class Tracker
 
       int state;
 
-      int servo_pos;
+      float servo_pos;
       int servo_direction;
       // the threshold delta voltage in which the tracker will determine whether or not it is hitting a reflective object
       int last_reading;
@@ -43,7 +43,7 @@ class Tracker
     static void execute_wrapper(void *self);
     int make_reading(Sensor *sensor);
     void move_servo(Sensor *sensor, int direction);
-    void move_servo(Sensor *sensor, int direction, int amount);
+    void move_servo(Sensor *sensor, int direction, float amount);
     void reverse_servo(Sensor *sensor);
 
     // determines whether there is one laser tracking one edge or two lasers tracking both
